Contact between halves of polygon
In the process of creating a contact between the two halves of the polygon.
Creating an iterator that uses the getclosest function and the array of generated vertices to get all the edges, figure out if they belong to the internal or external surface, and then generate sets accordingly. The two sets of interest are the side of the left polygon facing the right one and the side of the right polygon facing the left one. These are the two interacting sides that are most likely to engage each other during compression.